chore: rename `umbra_core` to `umbra`
