+++ packages/umbra/lib/src/generators/raw_generator.dart
@@ -1,7 +1,7 @@
 import 'package:mason/mason.dart';
-import 'package:umbra_core/src/generators/in_memory_generator_target.dart';
-import 'package:umbra_core/src/generators/templates/raw_file_bundle.dart';
-import 'package:umbra_core/umbra_core.dart';
+import 'package:umbra/src/generators/in_memory_generator_target.dart';
+import 'package:umbra/src/generators/templates/raw_file_bundle.dart';
+import 'package:umbra/umbra.dart';
 
 /// {@template raw_generator}
 /// Generates a raw GLSL shader from a [ShaderSpecification].
