+++ packages/umbra/lib/src/generators/templates/dart_file_bundle.dart
