+++ packages/umbra/lib/src/generators/templates/raw_file_bundle.dart
