+++ packages/umbra/lib/umbra_core.dart
@@ -1,4 +1,4 @@
-library umbra_core;
+library umbra;
 
 export 'src/generators/generators.dart';
 export 'src/models/models.dart';
