+++ packages/umbra/test/src/models/shader_specification_test.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
-import 'package:umbra_core/umbra_core.dart';
+import 'package:umbra/umbra.dart';
 
 import '../../helpers/helpers.dart';
