+++ packages/umbra_cli/lib/src/commands/generate/commands/spirv.dart
@@ -1,10 +1,10 @@
 import 'dart:io';
 
 import 'package:mason/mason.dart' hide packageVersion;
+import 'package:umbra/umbra.dart';
 import 'package:umbra_cli/src/commands/generate/commands/base_generate_command.dart';
 import 'package:umbra_cli/src/exit_with.dart';
 import 'package:umbra_cli/src/generators/generators.dart';
-import 'package:umbra_core/umbra_core.dart';
 
 /// {@template spirv_command}
 /// Generate a SPIR-V binary file based on the given shader file.
