+++ packages/umbra_cli/pubspec.yaml
@@ -13,8 +13,8 @@ dependencies:
   meta: ^1.7.0
   path: ^1.8.1
   pub_updater: ^0.2.2
-  umbra_core:
-    path: ../umbra_core
+  umbra:
+    path: ../umbra
 
 dev_dependencies:
   coverage: ^1.1.0
