+++ packages/umbra_cli/test/src/commands/generate/commands/base_command_generator_test.dart
@@ -6,9 +6,9 @@ import 'package:mason/mason.dart' hide packageVersion;
 import 'package:mocktail/mocktail.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
+import 'package:umbra/umbra.dart';
 import 'package:umbra_cli/src/commands/generate/commands/base_generate_command.dart';
 import 'package:umbra_cli/src/exit_with.dart';
-import 'package:umbra_core/umbra_core.dart';
 
 import '../../../../helpers/helpers.dart';
