+++ packages/umbra_flutter/pubspec.yaml
@@ -9,8 +9,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  umbra_core:
-    path: ../umbra_core
+  umbra:
+    path: ../umbra
   vector_math: ^2.1.1
 
 dev_dependencies:
